pyproject.toml
[project]
name = "intent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "starlette",
    "uvicorn[standard]",
    "pyyaml",
    "jsonschema",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
